+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 ### Spring Core/Boot
 
 - Added autocompletion in properties files for all classes fields
+- Inspection `unresolved property key` supports Relaxed Binding
 - Fixed inspection for `@Retention` on Spring annotations. Supported meta-annotations
+- Fixed stackoverflow error while property calculation on some cases
+- Support @Scheduled annotation
 
 ## [241.1.1263] - 2024-04-17
 
@@ -28,7 +31,6 @@
 
 ### Spring Core/Boot
 
-- Inspection `unresolved property key` supports Relaxed Binding
 - Fix: Kotlin constructor with default values counts as autowired
 - Inspection: Warning: Spring @Value annotation string should start with "#{" "${" or resource prefixes